Небольшие корректировки

+++ src/lesson6/Cat.java
@@ -3,7 +3,7 @@ package lesson6;
 public class Cat extends Animal {
     private static int catCounter = 0;
 
-    public Cat(String name, int runLimit, int swimLimit, int jumpLimit) {
+    public Cat(String name, int runLimit, int swimLimit, double jumpLimit) {
         super(name, runLimit, swimLimit, jumpLimit);
         catCounter++;
     }
