+++ src/lesson6/Dog.java
@@ -3,7 +3,7 @@ package lesson6;
 public class Dog extends Animal {
     private static int dogCounter = 0;
 
-    public Dog(String name, int runLimit, int swimLimit, int jumpLimit) {
+    public Dog(String name, int runLimit, int swimLimit, double jumpLimit) {
         super(name, runLimit, swimLimit, jumpLimit);
         dogCounter++;
     }
